pyproject.toml
[project]
name = "marketing-agents"
version = "0.1.0"
description = "Claude AI agent fleet for full-funnel marketing automation"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "rich>=13.0",          # pretty terminal output for reports
    "typer>=0.12",         # CLI for running agents manually
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
    "ruff>=0.4",
]

[project.scripts]
marketing-agents = "pipeline.cli:app"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
